fix: Add UUID and TIMESTAMPTZ type support in row deserialization
- Add UUID type handling using uuid::Uuid for PostgreSQL UUID columns
- Add TIMESTAMPTZ/TIMESTAMP type handling using chrono::DateTime<Utc>
- Change products.price from DECIMAL to DOUBLE PRECISION to avoid
  tokio_postgres NUMERIC deserialization issue
- Fix e2e tests to use snake_case field names and correct price type

+++ db/init.sql
@@ -76,7 +76,7 @@ CREATE TABLE IF NOT EXISTS public.products (
     id          UUID        PRIMARY KEY DEFAULT uuid_generate_v4(),
     name        VARCHAR(255) NOT NULL,
     description TEXT,
-    price       DECIMAL(10, 2) NOT NULL DEFAULT 0.00,
+    price       DOUBLE PRECISION NOT NULL DEFAULT 0.00,
     stock       INT          NOT NULL DEFAULT 0,
     is_active   BOOLEAN      NOT NULL DEFAULT TRUE,
     created_at  TIMESTAMPTZ  NOT NULL DEFAULT NOW()
